Feat/user login

+++ src/domains/User/services/UserService.ts
@@ -2,6 +2,8 @@ import prisma from '../../../../config/client';
 import { User } from '@prisma/client';
 import bcrypt from 'bcrypt';
 import { QueryError } from '../../../../errors/QueryError';
+import { Roles } from '../../../middlewares/checkRole';
+import { PermissionError } from '../../../../errors/PermissionError';
 class UserService{
 
 	async encryptPassword(password: string){
@@ -41,7 +43,7 @@ class UserService{
 		return ler;
 	}
 
-	async update (body: User){
+	async update (body: User, currentUser: User){
 		if(body.name == ''){
 			throw new QueryError('O usuário precisa de um nome.');
 		}
@@ -50,6 +52,15 @@ class UserService{
 		}
 		if(body.password == ''){
 			throw new QueryError('O usuário precisa de uma senha.');
+		}
+		if (!await this.findByEmail(body.email)){
+			throw new QueryError('O usuário não existe');
+		}
+		if (currentUser.email != body.email &&	 currentUser.role != Roles.admin){
+			throw new PermissionError('Você não tem permissão para realizar essa ação.');
+		}
+		if (body.role && currentUser.role != Roles.admin){
+			throw new PermissionError('Você não tem permissão para realizar essa ação.');
 		}
 		const atualizar = await prisma.user.update({
 			where:{email: body.email},
@@ -67,7 +78,15 @@ class UserService{
 		return atualizar;
 	}
 
-	async delete (email: string){
+	async delete (email: string, currentUser:User){
+		const user = await this.findByEmail(email);
+		
+		if (!user){
+			throw new QueryError('O usuário não existe');
+		}
+		if (email != currentUser.email && currentUser.role != Roles.admin){
+			throw new PermissionError('Você não tem permissão para realizar essa ação');
+		}
 		const deletar = await prisma.user.delete({
 			where: {email: email}
 		});
